Add star_analyzer tool for STAR archive analysis and static analysis script
- Implemented star_analyzer.c for detailed analysis of STAR archives, including:
  - Basic, detailed, compression, integrity, and optimization analysis modes
  - Comprehensive statistics collection (file counts, size statistics, compression ratios, symbol statistics)
  - Output options for JSON and CSV formats
  - Command line argument parsing and usage instructions

- Created static_analysis.sh for comprehensive static analysis using tools like cppcheck and clang-static-analyzer:
  - Configurable options for cleaning previous results, selecting analysis tools, and output formats
  - Dependency checks for required analysis tools
  - Functions for running cppcheck, clang static analyzer, splint, PC-lint, and Valgrind
  - Summary report generation and CI-friendly output

+++ tests/unity/unity.c
@@ -228,9 +228,9 @@ void unity_test_assert_equal_hex32(const uint32_t expected,
     if (expected != actual) {
         Unity.CurrentTestFailed = 1;
         unity_print("Expected ");
-        unity_print_number(expected, UNITY_DISPLAY_STYLE_HEX32);
+        unity_print_number((int32_t)expected, UNITY_DISPLAY_STYLE_HEX32);
         unity_print(" Was ");
-        unity_print_number(actual, UNITY_DISPLAY_STYLE_HEX32);
+        unity_print_number((int32_t)actual, UNITY_DISPLAY_STYLE_HEX32);
         
         if (msg != NULL) {
             unity_print(" - ");
